app/requirements.txt
customtkinter>=5, <6
tkinterdnd2>=0.4, <0.5
Pillow>=11, <12
tensorflow~=2.17.0
numpy>=1.26, <2
